extract common utility functions
